Refactor: BottomSheetView 분리 (#12)
